requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.3.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
narwhals==2.4.0
numpy==2.2.6
packaging==25.0
pandas==2.3.2
pillow==11.3.0
postgrest==1.1.1
protobuf==6.32.0
psycopg2-binary==2.9.10
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
realtime==2.7.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
storage3==0.12.1
streamlit==1.49.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
